+++ app/container.tsx
@@ -1,5 +1,4 @@
-import Game from '@components/game/game';
-import Highlights from '@components/Highlights';
+import CodeShowcase from '@components/code-showcase/CodeShowcase';
 
 interface Props {
   children: React.ReactNode;
@@ -7,14 +6,11 @@ interface Props {
 
 export default function HomePageContainer({ children }: Props) {
   return (
-    <main className="flex flex-col gap-y-6 p-4">
-      <div className="flex flex-col lg:flex-row gap-6 items-center justify-between w-full max-w-screen-lg mx-auto">
-        <div id="intro" className="flex-1 rounded-lg p-6 w-full">
-          {children}
-        </div>
-        <Game />
+    <div className="flex flex-col lg:flex-row gap-6 items-center justify-between w-full max-w-screen-lg mx-auto min-h-[95%]">
+      <div id="intro" className="flex-1 rounded-lg p-6 w-full flex-col items-center">
+        {children}
       </div>
-      <Highlights />
-    </main>
+      <CodeShowcase />
+    </div>
   );
 }